requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
easyocr
regex
Pillow
bcrypt